third_party/blink/web_tests/platform/win11-arm64/http/tests/media/track/track-webvtt-slow-loading-2-expected.txt

This is a testharness.js-based test.
FAIL Slow loading WebVTT file interrupted in the middle of a timestamp line. assert_equals: expected 2 but got 0
Harness: the test ran to completion.
